pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-persona-sim"
version = "0.1.0"
description = "Persona and memory simulation with chat and decision modes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.0.0",
  "faiss-cpu>=1.8.0",
  "numpy>=1.26.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.0",
  "typer>=0.12.0"
]

[project.scripts]
persona-sim = "ai_persona_sim.cli:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
